refactor(docs): use only route.module.Sidebar for website
